pyproject.toml
[tool.poetry]
name = "pangeo-forge-bakery-images"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.102.0"
"aws-cdk.aws-ecr" = "^1.102.0"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.2"
isort = "^5.8.0"
python-dotenv = "^0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
